ex 4
