+++ modulo005/customer/src/main/java/com/globallogic/bootcampms/customer/models/Customer.java
@@ -1,5 +1,7 @@
 package com.globallogic.bootcampms.customer.models;
 
+import org.springframework.beans.factory.annotation.Value;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,9 +12,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Customer {
+	
+	@Value("${customer.id}")
 	private Long id;
+	
+	@Value("${customer.dni}")
 	private Long dni;
+	
+	@Value("${customer.name}")
 	private String nombre;
+	
+	@Value("${customer.apellido}")
 	private String apellido;
 	
 	@Override
