+++ modulo005/customer/src/main/resources/application.properties
@@ -1,1 +1,4 @@
-
+customer.name="José"
+customer.apellido="Pérez"
+customer.id=12342
+customer.dni=234567890
