--- modulo005/order/src/main/java/com/globallogic/bootcampgl/order/utils/OrderFactory.java
@@ -1,13 +1,0 @@
-package com.globallogic.bootcampgl.order.utils;
-
-import com.globallogic.bootcampgl.order.models.Order;
-
-public class OrderFactory {
-
-	public static Order getNewOrder() {
-		return Order.builder()
-				.productName("C-00221").count(1000l).customerId(12342l)
-				.build();
-	}
-	
-}
